Moved components in their respective directories. TODO: fetch product data from db
